+++ Navigation/NXE/src/navitipc.h
@@ -10,7 +10,7 @@ public:
     //! IPC start/stop
     virtual void start() = 0;
     virtual void stop() = 0;
-    virtual void moveBy(double x, double y) = 0;
+    virtual void moveBy(int x, int y) = 0;
     virtual int zoom() = 0;
     virtual void zoomBy(int factor) = 0;
     virtual void render() = 0;
