+++ Navigation/NXE/tests/unit/mocks/navitcontrollermock.h
@@ -7,7 +7,7 @@ struct NavitIPCMock : public NXE::NavitIPCInterface {
     GMOCK_METHOD0_(, , , start, void());
     GMOCK_METHOD0_(, , , stop, void());
     GMOCK_METHOD0_(, , , render, void());
-    GMOCK_METHOD2_(, , , moveBy, void(double, double));
+    GMOCK_METHOD2_(, , , moveBy, void(int, int));
     GMOCK_METHOD1_(, , , zoomBy, void(int));
     GMOCK_METHOD0_(, , , zoom, int(void));
 };
